[IMPROVEMENT] Use react-native-notifier for in-app notifications (#2139)
