+++ app/lib/methods/subscriptions/rooms.js
@@ -9,12 +9,12 @@ import log from '../../../utils/log';
 import random from '../../../utils/random';
 import store from '../../createStore';
 import { roomsRequest } from '../../../actions/rooms';
-import { notificationReceived } from '../../../actions/notification';
 import { handlePayloadUserInteraction } from '../actions';
 import buildMessage from '../helpers/buildMessage';
 import RocketChat from '../../rocketchat';
-import EventEmmiter from '../../../utils/events';
+import EventEmitter from '../../../utils/events';
 import { removedRoom } from '../../../actions/room';
+import { INAPP_NOTIFICATION_EMITTER } from '../../../containers/InAppNotification';
 
 const removeListener = listener => listener.stop();
 
@@ -267,7 +267,7 @@ export default function subscribeRooms() {
 					if (data.rid === roomState.rid && roomState.isDeleting) {
 						store.dispatch(removedRoom());
 					} else {
-						EventEmmiter.emit('ROOM_REMOVED', { rid: data.rid });
+						EventEmitter.emit('ROOM_REMOVED', { rid: data.rid });
 					}
 				} catch (e) {
 					log(e);
@@ -320,7 +320,7 @@ export default function subscribeRooms() {
 			} catch (e) {
 				// do nothing
 			}
-			store.dispatch(notificationReceived(notification));
+			EventEmitter.emit(INAPP_NOTIFICATION_EMITTER, notification);
 		}
 		if (/uiInteraction/.test(ev)) {
 			const { type: eventType, ...args } = type;
