--- app/reducers/notification.js
@@ -1,24 +1,0 @@
-import { NOTIFICATION } from '../actions/actionsTypes';
-
-const initialState = {
-	message: '',
-	payload: {
-		type: 'p',
-		name: '',
-		rid: ''
-	}
-};
-
-export default function notification(state = initialState, action) {
-	switch (action.type) {
-		case NOTIFICATION.RECEIVED:
-			return {
-				...state,
-				...action.payload
-			};
-		case NOTIFICATION.REMOVE:
-			return initialState;
-		default:
-			return state;
-	}
-}
