+++ package.json
@@ -83,6 +83,7 @@
     "react-native-modal": "11.5.6",
     "react-native-navigation-bar-color": "2.0.1",
     "react-native-notifications": "2.1.7",
+    "react-native-notifier": "^1.3.1",
     "react-native-orientation-locker": "1.1.8",
     "react-native-picker-select": "7.0.0",
     "react-native-platform-touchable": "^1.1.1",
